requirements.txt
opencv-python
numpy
matplotlib
tensorflow
pillow
PyMuPDF
gradio
scikit-learn
